Fixed ArchGDal, NetCDF, Zarr further
